feat: youtube category slice 이동 및 적용
