+++ src/features/influencer/index.ts
@@ -7,19 +7,16 @@ export { OutlierRangeDropdown } from './ui/OutlierRangeDropdown'
 export { HasAdHistoryDropdown } from './ui/HasAdHistoryDropdown'
 export { EngagementRateDropdown } from './ui/EngagementRateDropdown'
 export {
-  fetchYoutubeCategories,
   fetchInfluencers,
   addBookmark,
   removeBookmark,
 } from './api/influencerApi'
-export { useInfluencers, useYoutubeCategories } from './model/useInfluencers'
+export { useInfluencers } from './model/useInfluencers'
 export type {
-  YoutubeCategoriesResponse,
   InfluencerListResponse,
   BookmarkResponse,
   FetchInfluencersParams,
 } from './api/influencerApi'
-export type { YoutubeCategory } from './mock/mockYoutubeCategories'
 export type { SubscriberQuery } from './ui/SubscriberDropdown'
 export type {
   EngagementRateQuery,
